Se vuelve actualizar y se termina CU-02: Registrar Productos

+++ src/main/java/co/uncetral/sistemas/AppInsumos/repositorios/RepositorioProductos.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RepositorioProductos extends JpaRepository<ProductosEntidades,Integer> {
+public interface RepositorioProductos extends JpaRepository<ProductosEntidades,Long> {
 }
